% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.thresh.R
\name{addExcesses}
\alias{addExcesses}
\title{Annotate a threshold selection ggplot}
\usage{
addExcesses(p, x, y, data, textsize)
}
\arguments{
\item{p}{An object produced by ggplot}

\item{x}{Horizontal axis data containing the full range.}

\item{y}{Verticle axis data containing the full range.}

\item{data}{The actual data being considered for GPD modelling.}

\item{textsize}{The size of the text in the annotations.}
}
\description{
Annotate a threshold selection ggplot with the number
of exceedances of various thresholds.
}
