% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.evm.R
\name{AIC.evmOpt}
\alias{AIC.evmOpt}
\title{Akaike's Information Criterion}
\usage{
\method{AIC}{evmOpt}(object, penalized = FALSE, ..., k = 2)
}
\arguments{
\item{object}{fit model object}

\item{penalized}{whether to use the penalized log-likelihood}

\item{...}{other arguments currently ignored}

\item{k}{numeric, the penalty per parameter to be used; the
default \code{k = 2} is the classical AIC.}
}
\value{
the AIC
}
\description{
Compute AIC for \code{evmOpt} objects
}
\seealso{
\code{\link[stats]{AIC}}
}

