% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmBoot.R
\name{coef.evmBoot}
\alias{coef.evmBoot}
\title{Extract coefficients from an evmBoot fit}
\usage{
\method{coef}{evmBoot}(object, ...)
}
\arguments{
\item{object}{a \code{\link{evmBoot}} object}

\item{...}{ignored}
}
\value{
the mean value of each coefficient
}
\description{
Extract coefficients from an evmBoot fit
}

