% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.evmOpt}
\alias{cv.evmOpt}
\title{Cross-validation for the shape parameter in an extreme values model}
\usage{
\method{cv}{evmOpt}(object, folds = 10, ..., penalty = "gaussian",
  range = seq(1, 25, length.out = 25), shape = NULL)
}
\arguments{
\item{object}{An object of class 'evmOpt' as returned by \code{evm}.}

\item{folds}{Integer giving the number of cross-validation folds to use.
Defaults to \code{folds = 10}.}

\item{...}{Not used.}

\item{penalty}{String specifying the type of penalty to use. Defaults to
\code{penalty = "gaussian"} which is equivalent to using a quadratic
penalty. The other allowed value is \code{penalty = "lasso"} and an L1
penalty is used.}

\item{range}{A sequence of values for the penalty parameter. Defaults to
\code{range = seq(1, 25, length.out = 25)}. Must be strictly positive.
The values are taken to be the reciprocals of the prior variance so mush
be strictly positive.}

\item{shape}{String giving the name of the shape parameter. Defaults to
\code{shape = NULL} and the function tries to guess.}
}
\description{
Cross-validation for the shape parameter in an extreme values model
}
\details{
Only the shape parameter is assumed to be penalized. The penalty
  can be thought of in terms of the variance of a prior distribution, which
  is equivalent to a quadratic penalty. Because the shape parameter will
  usually be between -1/2 and 1/2, a prior N(0, 1/16) distribution will
  likely be a good starting point, so values that span 16 will usually be
  appropriate.
}
