% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glo.R
\name{rglo}
\alias{rglo}
\alias{pglo}
\alias{qglo}
\alias{dglo}
\title{Generalized logistic distribution}
\usage{
rglo(n, mu, sigma, xi)

dglo(x, mu, sigma, xi, log.d = FALSE)

pglo(q, mu, sigma, xi, lower.tail = TRUE, log.p = FALSE)

qglo(p, mu, sigma, xi, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{Number of random numbers to generate.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{xi}{Shape parameter.}

\item{x, q, p}{Value, quantile or probability respectively.}

\item{log.d, log.p}{Whether to work on the log scale.}

\item{lower.tail}{Whether to return the lower tail.}
}
\description{
Density, distribution and quantile functions, and random number
  generation for the Generalized logistic distribution
}
\concept{pglo qglo dglo rglo}
