% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.exprel}
\alias{.exprel}
\title{Accurately compute (exp(x) - 1) / x}
\usage{
.exprel(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Accurately compute (exp(x) - 1) / x
}
