% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.log1mexp}
\alias{.log1mexp}
\title{Accurately compute log(1-exp(x))}
\usage{
.log1mexp(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a numeric vector
}
\description{
Accurately compute log(1-exp(x))
}
