% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.log1prel}
\alias{.log1prel}
\title{Accurately compute log(1 + x) / x}
\usage{
.log1prel(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Accurately compute log(1 + x) / x
}
