% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.specfun.safe.product}
\alias{.specfun.safe.product}
\title{Compute pmax(x y, -1) in such a way that zeros in x beat
infinities in y.}
\usage{
.specfun.safe.product(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
an appropriate numeric vector
}
\description{
This is a common pattern in much of the distribution code, so it's
worth factoring out.
}
