% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.cluster.R
\name{ggplot.declustered}
\alias{ggplot.declustered}
\alias{ggplot.extremalIndex}
\title{Diagnostic plots for an declustered object}
\usage{
\method{ggplot}{declustered}(
  data = NULL,
  mapping,
  xlab,
  ylab,
  main,
  ptcol = c("blue", "orange"),
  col = "light blue",
  plot. = TRUE,
  ...,
  environment
)

\method{ggplot}{extremalIndex}(
  data = NULL,
  mapping,
  xlab,
  ylab,
  main,
  ptcol = "blue",
  col = "light blue",
  plot. = TRUE,
  ...,
  environment
)
}
\arguments{
\item{data}{An object of class \code{declustered} or \code{extremalIndex}.}

\item{mapping}{Not used.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{main}{Plot title.}

\item{ptcol}{Colour for points. Defaults to \code{ptcol="blue"}.}

\item{col}{Colour for lines. Defaults to \code{col="light blue"}.}

\item{plot.}{Whether or not to display the output. Defaults to \code{plot.=TRUE}.}

\item{...}{Other arguments passed through to underlying plot functions.}

\item{environment}{Not used.}
}
\description{
Create and display diagnostic plots for a declustered object.
}
