% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.predict.R
\name{ggplot.rl.evmOpt}
\alias{ggplot.rl.evmOpt}
\alias{ggplot.rl.evmSim}
\alias{ggplot.rl.evmBoot}
\alias{ggplot.lp.evmOpt}
\alias{ggplot.lp.evmSim}
\alias{ggplot.lp.evmBoot}
\title{Plotting function for return level estimation}
\usage{
\method{ggplot}{rl.evmOpt}(
  data = NULL,
  mapping,
  xlab,
  ylab,
  main,
  ylim = "auto",
  ptcol = "blue",
  col = "light blue",
  fill = "orange",
  alpha = 0.5,
  ...,
  environment
)
}
\arguments{
\item{data}{An object of class \code{rl.evmOpt}, \code{rl.evmBoot}, \code{rl.evmSim}, \code{lp.evmOpt}, \code{lp.evmBoot} or \code{lp.evmSim}, .}

\item{mapping}{Not used.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{main}{Plot title.}

\item{ylim}{Plot limits for y-axis.}

\item{ptcol}{Colour for points. Defaults to \code{ptcol="blue"}.}

\item{col}{Colour for lines. Defaults to \code{col="light blue"}.}

\item{fill}{Colour for shading polygons.}

\item{alpha}{Transparency.}

\item{...}{Other arguments passed through to underlying plot functions.}

\item{environment}{Not used.}
}
\description{
Plotting function for return level estimation
}
