% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.evm.R
\name{logLik.evmOpt}
\alias{logLik.evmOpt}
\title{Log-likelihood for evmOpt objects}
\usage{
\method{logLik}{evmOpt}(object, penalized = FALSE, ...)
}
\arguments{
\item{object}{fit model object}

\item{penalized}{whether to return the penalized log-likelihood}

\item{...}{some methods need more arguments}
}
\value{
an object of class \code{logLik}
}
\description{
Return the log-likelihood or penalized log-likelihood for
\code{evmOpt} objects.
}
\seealso{
\code{\link[stats]{logLik}}
}
