% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmBoot.R
\name{plot.evmBoot}
\alias{plot.evmBoot}
\title{Plot an evmBoot object}
\usage{
\method{plot}{evmBoot}(x, col = 4, border = NULL, ...)
}
\arguments{
\item{x}{an \code{\link{evmBoot}} object}

\item{col}{colour used to fill histogram}

\item{border}{the colour of the border around the bars}

\item{...}{other arguments passed to \code{\link[graphics]{hist}}}
}
\description{
Plot the bootstrap distribution of the model parameters.
}

