\name{print.cmvxDependence}
\alias{print.cmvxDependence}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print the parameters in a multivariate conditional dependence extreme value model. }
\description{
Show dependence structure parameters from an object of class \code{cmvxDependence}.
}
\usage{
print.cmvxDependence(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ An object of class \code{cmvxDependence}. }
}
\details{
    The parameters in the dependence structure of a multivariate extreme value
    model, as described by Heffernan and Tawn, 2004.
}
\value{NULL}
\references{ J. E. Heffernan and J. A. Tawn, A conditional approach for multivariate
extreme values, Journal of the Royal Statistical Society (B), 66, 497 -- 546, 2004. }
\author{ Harry Southworth }
\seealso{ \code{\link{migpd}}, \code{\link{cmvxBoot}}, \code{\link{cmvxPrediction}} }
\examples{

}
\keyword{ models }
\keyword{ multivariate }
