% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.evm.R
\name{print.evmOpt}
\alias{print.evmOpt}
\title{Print evmOpt objects}
\usage{
\method{print}{evmOpt}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{a fit evmOpt object}

\item{digits}{number of digits used for printing}

\item{...}{further arguments passed to \code{\link{format}}}
}
\description{
Print evmOpt objects
}
