\name{rain, wavesurge and portpirie}
\alias{rain}
\alias{wavesurge}
\alias{portpirie}
\docType{data}
\title{
Rain, wavesurge and portpirie datasets.
}
\description{
Rainfall, wave-surge and Port Pirie data sets.
}
\usage{
data(rain)
data(wavesurge)
data(portpirie)
}
\format{
 The format of the rain data is:
 num [1:17531] 0 2.3 1.3 6.9 4.6 0 1 1.5 1.8 1.8 ...

  The wave-surge data is bivariate and is used for testing
  functions in \code{texmex}.

  The Port Pirie data has two columns: 'Year' and 'SeaLevel'.
}
\details{
The datasets are used by Coles and appear in the \code{ismev} package.
}
\source{
Copied from the \code{ismev} package.
}
\references{
S. Coles, An Introduction to Statistical Modeling of Extreme Values, Springer, 2001
}
\keyword{datasets}
