% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMaxAR.R
\name{rFrechet}
\alias{rFrechet}
\title{Extreme Value random process generation.}
\usage{
rFrechet(n)
}
\arguments{
\item{n}{Number of samples to generate.}
}
\description{
Extreme Value random process generation.
}
\details{
Generation of samples from unit Frechet processes.
}
\examples{
rFrechet(1000)
}
