% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glo.R
\name{rglo}
\alias{rglo}
\alias{pglo}
\alias{qglo}
\alias{dglo}
\title{Generalized logistic distribution}
\usage{
rglo(n, mu, sigma, xi)
}
\arguments{
\item{n}{Number of random numbers to generate.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{xi}{Shape parameter.}

\item{x, q, p}{Value, quantile or probability respectively.}

\item{log.d, log.p}{Whether to work on the log scale.}

\item{lower.tail}{Whether to return the lower tail.}
}
\description{
Generalized logistic distribution
}
