% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMaxAR.R
\name{rMaxAR}
\alias{rMaxAR}
\title{Extreme Value random process generation.}
\usage{
rMaxAR(n,theta)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{theta}{Parameter of the MAX AR process, takes values between 0 and 1.}
}
\description{
Extreme Value random process generation.
}
\details{
Generation of samples from  Max AR(theta) processes.
}
\examples{
rMaxAR(1000,0.2)
}
