% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\alias{print.cv}
\alias{summary.cv}
\alias{plot.cv}
\alias{ggplot.cv}
\title{Cross-validation for a model object}
\usage{
cv(object, folds = 10, ...)

\method{print}{cv}(x, ...)

\method{summary}{cv}(object, ...)

\method{plot}{cv}(x, y, ...)

\method{ggplot}{cv}(data, mapping = NULL, ..., environment = parent.frame())
}
\arguments{
\item{object}{A model object.}

\item{folds}{The number of cross-validation folds to use. Defaults to
\code{folds = 10}.}

\item{...}{Other arguments to be passed through to methods.}

\item{x, y}{Arguments to plot method.}

\item{data, mapping, environment}{Arguments ggplot method.}
}
\description{
Cross-validation for a model object
}
\details{
The function is generic. At present, only objects of class 'evmOpt',
  as returned by \code{texmex::evm} can be used.
}
\seealso{
\code{\link{cv.evmOpt}}
}
