% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{dgumbel}
\alias{dgumbel}
\alias{rgumbel}
\alias{pgumbel}
\alias{qgumbel}
\title{The Gumbel distribution}
\usage{
dgumbel(x, mu, sigma, log.d = FALSE)

rgumbel(n, mu, sigma)

pgumbel(q, mu, sigma, lower.tail = TRUE, log.p = FALSE)

qgumbel(p, mu, sigma, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x, q, p}{Vectors of quantiles or probabilities.}

\item{mu}{The location parameter.}

\item{sigma}{The scale parameter.}

\item{log.d, log.p}{Whether to return logged values, or to treat probabilities/densities as being logged.}

\item{n}{The number of observations.}

\item{lower.tail}{Whether to return the lower tail. If \code{lower.tail=FALSE},
the upper tail is returned.}
}
\description{
Density, distribution and quantile functions, and random number
  generation for the Gumbel distribution
}
\concept{rgumbel pgumbel qgumbel dgumbel}
