% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulas.R
\name{edf}
\alias{edf}
\title{Compute empirical distribution function}
\usage{
edf(x, na.last = NA)
}
\arguments{
\item{x}{A numeric vector}

\item{na.last}{How to treat missing values. See \code{\link{rank}} for
details.}
}
\value{
A vector of quantiles relating to the observations in \code{x}.
}
\description{
Compute the empirical distribution function
}
\examples{

plot(winter$NO, edf(winter$NO))

}
\seealso{
\code{\link{copula}}
}
\author{
Harry Southworth
}
\keyword{univar}
