% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.bayes.R
\name{ggplot.evmSim}
\alias{ggplot.evmSim}
\alias{ggtraceplots}
\alias{ggdensplots}
\alias{ggacfplots}
\title{Diagnostic plots for the Markov chains in an evmSim object}
\usage{
\method{ggplot}{evmSim}(
  data = NULL,
  mapping,
  which.plots = 1:3,
  chain = 1,
  denscol = "dark blue",
  acfcol = "light blue",
  plot.it = TRUE,
  ...,
  environment
)
}
\arguments{
\item{data}{An object of class 'evmSim'.}

\item{mapping, environment}{ignored}

\item{which.plots}{Which plots to produce. Density plots correspond
to 1, trace plots of the Markov chains to 2 and autocorrelation
function plots to 3.}

\item{chain}{An integer indicating which chain to plot (only relevant if there
is more than 1 chain). Defaults to 1. If you ran multiple chains, you
should look at diagnostics for all of them.}

\item{denscol}{Colour for the density plots. Defaults to 'dark blue'.}

\item{acfcol}{Colour for the ACF plots. Defaults to 'light blue'.}

\item{plot.it}{Whether or not to actually print the plots. Defaults
to \code{plot.it=TRUE}.  If \code{plot.it=FALSE}, you might
want to control the layout. Do this with
\code{do.call("grid.arrange", c(plots, ncol=2))}, for example,
where \code{plots} is the objected returned by
\code{ggplot.evmSim}.}

\item{...}{Additional arguments to \code{ggplot}, currently unused.}
}
\description{
Diagnostic plots for the Markov chains in an evmSim object
}
\keyword{hplot}
