% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulas.R
\name{plot.copula}
\alias{plot.copula}
\title{Plot copulas}
\usage{
\method{plot}{copula}(x, jitter. = FALSE, jitter.factor = 1, ...)
}
\arguments{
\item{x}{A copula object}

\item{jitter.}{If \code{jitter=TRUE}, the values are jittered
before plotting. Defaults to \code{jitter. = FALSE}.}

\item{jitter.factor}{How much jittering to use. Defaults to
\code{jitter.factor = 1}.}

\item{...}{Other arguments to pass through to \code{plot}.}
}
\description{
Plot copulas
}
