\name{rain and wavesurge}
\alias{rain}
\alias{wavesurge}
\docType{data}
\title{
Rain and wavesurge datasets.
}
\description{
Rainfall and wave-surge data sets.
}
\usage{data(rain)
data(wavesurge)
}
\format{
  The format of the rain data is:
 num [1:17531] 0 2.3 1.3 6.9 4.6 0 1 1.5 1.8 1.8 ...

  The wave-surge data is biviate and is used for testing
functions in \code{texmex}.
}
\details{
The datasets are used by Coles and appear in the \code{ismev} package.
}
\source{
Copied from the \code{ismev} package.
}
\references{
S. Coles, An Introduction to Statistical Modeling of Extreme Values, Springer, 2001
}
\keyword{datasets}
